refactor(cli): simplify BookCreateCommands by removing unused dependencies and methods
