+++ src/main/java/com/penrose/bibby/cli/command/book/BookCreateScanCommands.java
@@ -1,13 +1,11 @@
-package com.penrose.bibby.cli.command.book.create;
+package com.penrose.bibby.cli.command.book;
 
-import com.penrose.bibby.cli.command.book.ScanMode;
 import com.penrose.bibby.cli.prompt.application.CliPromptService;
 import com.penrose.bibby.cli.prompt.domain.PromptOptions;
 import com.penrose.bibby.cli.ui.BookcardRenderer;
 import com.penrose.bibby.library.cataloging.author.contracts.AuthorDTO;
 import com.penrose.bibby.library.cataloging.author.contracts.ports.AuthorFacade;
 import com.penrose.bibby.library.cataloging.book.contracts.dtos.BookMetaDataResponse;
-import com.penrose.bibby.library.cataloging.book.contracts.dtos.BookRequestDTO;
 import com.penrose.bibby.library.cataloging.book.contracts.ports.inbound.BookFacade;
 import com.penrose.bibby.library.stacks.bookcase.contracts.ports.inbound.BookcaseFacade;
 import com.penrose.bibby.library.stacks.shelf.contracts.ports.inbound.ShelfFacade;
